refactor(cognitive_memory): split mod.rs into backup + ops (#1266)
cognitive_memory/mod.rs was 1014 LOC. Split:
- mod.rs (318): NativeCognitiveMemory struct + open/in_memory/open_read_only +
  conn/execute/ensure_schema/new_id/now_secs/as_str/as_i64/as_f64
- backup.rs (396): WAL paths + DB recovery + verify + restore + create/prune backups + with_open_lock
- ops.rs (328): impl CognitiveMemoryOps + escape_cypher

All 21 cognitive_memory tests pass.
